Implement basic NOT
We process each part of an AND query in the same way, but after
processing we check if it is a NOT query. If it is, then we use the
basic NOT function to invert the result.

This so far, does not use the advanced NOT operators like AND_NOT or
OR_NOT.